refactor(jit): extract x86_64 backend into separate file
